fixed most bugs + output

+++ MemoryManager.cpp
@@ -27,7 +27,7 @@ void MemoryManager::initAllocator(size_t max_overall_mem, size_t mem_per_frame)
   deleteFileInfo();
 }
 
- size_t MemoryManager::getMemUsed() const {
+ uint64_t MemoryManager::getMemUsed() const {
    if (mem_allocator) {
      return mem_allocator->getMem();
    }
